requirements.txt
discord.py==2.2.3
asqlite @ git+https://github.com/Rapptz/asqlite.git@ed54062
